requirements.txt
alembic==1.6.5
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
click==8.0.1
colorama==0.4.4
distlib==0.3.2
fastapi==0.68.0
filelock==3.0.12
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
platformdirs==2.2.0
postgres==3.0.0
psycopg2-binary==2.9.1
psycopg2-pool==1.1
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
pytube==10.9.3
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.22
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
virtualenv==20.6.0
watchgod==0.7
websockets==9.1
Werkzeug==2.0.1
